Broke out network utillities. Needs testing.
