+++ miscutils.cc
@@ -1,0 +1,9 @@
+#include "miscutils.h"
+#include <iostream>
+#include <cstring>
+
+void errorexit(const char* message)
+{
+    std::cerr << message << ": " << strerror(errno) << std::endl;
+    exit(errno);
+}
